build: fix clang build

+++ src/curves/driver/math64.h
@@ -50,7 +50,7 @@ static inline int64_t __div_s128_by_s64(int128_t dividend, int64_t divisor)
 // Generic case: Use compiler's existing 128-bit division operator.
 static inline int64_t __div_s128_by_s64(int128_t numerator, int64_t denominator)
 {
-	return numerator / denominator;
+	return (int64_t)(numerator / denominator);
 }
 
 #endif
